refactored dom and app logic
